实现用户的添加和修改，sql修改

+++ application/sql/v1.0.0.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
   `id` int(8) NOT NULL AUTO_INCREMENT COMMENT '唯一id,主键',
   `name` varchar(128) COLLATE utf8_unicode_ci NOT NULL COMMENT '名称',
-  `job_number` tinyint(6) NOT NULL COMMENT '工号',
+  `job_number` int(6) NOT NULL COMMENT '工号',
   `depart_number` tinyint(2) NOT NULL COMMENT '部门编号',
   `password` varchar(128) COLLATE utf8_unicode_ci NOT NULL COMMENT '密码',
   `email` varchar(128) COLLATE utf8_unicode_ci DEFAULT NULL COMMENT '邮箱',
